requirements.txt
streamlit
torch
transformers
diffusers
accelerate
safetensors
sentencepiece
python-docx
bitsandbytes
